+++ src/biv_lite/biv_frames.py
@@ -6,13 +6,17 @@ import re
 
 class BivFrames(List[BivMesh]):
     """
-    A class that handles multiple BivMesh frames as a full cardiac cycle.
+    This simply a list of BivMesh objects with some extra functions specific for cardiac motion.
     """
     @classmethod
-    def from_folder(cls, folder: Path, pattern: str = "*_Model_Frame_*.txt", frame_str: str = r'_(\d+).txt'):
+    def from_folder(cls, folder: Path | str, pattern: str = "*_Model_Frame_*.txt", frame_str: str = r'_(\d+).txt'):
         biv = BivFrames()
 
-        for i, input_file in enumerate(sorted(folder.glob(pattern), key=lambda p: int(re.search(frame_str, p.name).groups()[0]))):
+        for i, input_file in enumerate(sorted(Path(folder).glob(pattern), key=lambda p: int(re.search(frame_str, p.name).groups()[0]))):
             biv.append(BivMesh.from_fitted_model(input_file, name=f"frame_{i}"))
 
         return biv
+
+    def gls(self):
+        """Compute global longitudinal strain values."""
+        # TODO: complete this by copying codes frm biv-me's calculate_longitudinal_strain